fix: give Video compilation a shared owner outside project storage
