+++ server/lib/index.js
@@ -587,3 +587,4 @@ export * from './privateSecuritySandbox.js';
 export * from './requestOrigin.js';
 export * from './cosFederationPolicy.js';
 export * from './creativeDirectorVideoReview.js';
+export * from './creativeDirectorVideoCompiler.js';
