+++ server/services/creativeDirector/projectsLogic.js
@@ -11,6 +11,8 @@
  */
 
 import { videoSceneInputs, retainVideoCuts } from '../../lib/creativeDirectorVideoReview.js';
+import { compileVideoArtifact, validateVideoShot } from '../../lib/creativeDirectorVideoCompiler.js';
+
 import { randomUUID } from 'crypto';
 import { EFFORT_LEVELS } from '../../lib/providerModels.js';
 import { ServerError } from '../../lib/errorHandler.js';
@@ -21,8 +23,9 @@ import { compareNewerWins } from '../../lib/lwwTimestamp.js';
 import { pickLlmRoutePinLayer } from '../../lib/llmRoutePin.js';
 import { localImageFilename } from '../../lib/localImageFilename.js';
 import { sanitizeProjectForSync } from '../../lib/projectStoreKit.js';
-import { GROK_VIDEO_DURATIONS } from '../../lib/grokVideoClip.js';
-import { REACTOR_MIN_CLIP_SECONDS, REACTOR_MAX_CLIP_SECONDS, REACTOR_MAX_PROMPT_LENGTH, REACTOR_ASPECTS } from '../../lib/reactorVideoClip.js';
+
+// Preserve the existing validation export for callers on the project-store surface.
+export { validateVideoShot } from '../../lib/creativeDirectorVideoCompiler.js';
 
 export { sanitizeProjectForSync } from '../../lib/projectStoreKit.js';
 
@@ -376,77 +379,6 @@ export function applyProjectPatch(project, patch) {
   return next;
 }
 
-export function validateVideoShot(project, scene, isFirst) {
-  const backend = project.renderBackend?.video?.mode;
-  const unsupported = [];
-  if (isFirst && scene.useContinuationFromPrior) unsupported.push('continuation without a prior shot');
-  if (backend === 'grok' && !GROK_VIDEO_DURATIONS.includes(scene.durationSeconds)) {
-    unsupported.push(`duration (choose ${GROK_VIDEO_DURATIONS.join(' or ')} seconds)`);
-  }
-  if (backend === 'reactor') {
-    if (scene.durationSeconds < REACTOR_MIN_CLIP_SECONDS || scene.durationSeconds > REACTOR_MAX_CLIP_SECONDS) {
-      unsupported.push(`duration (choose ${REACTOR_MIN_CLIP_SECONDS}–${REACTOR_MAX_CLIP_SECONDS} seconds)`);
-    }
-    if (scene.prompt.length > REACTOR_MAX_PROMPT_LENGTH) unsupported.push(`prompt (maximum ${REACTOR_MAX_PROMPT_LENGTH} characters)`);
-    if (!REACTOR_ASPECTS.includes(project.aspectRatio)) unsupported.push('aspect ratio');
-  }
-  if (unsupported.length) {
-    throw new ServerError(`Video shot ${scene.sceneId} has incompatible ${unsupported.join(', ')}. Revise the shot or choose a compatible backend.`, {
-      status: 400, code: 'VIDEO_BACKEND_INPUT_UNSUPPORTED',
-    });
-  }
-}
-
-/** Compile the existing treatment into a persisted Video artifact, without dispatch. */
-function compileVideoArtifact(project, treatment, sourceRevisions) {
-  if (!treatment.script) {
-    throw new ServerError('Video treatments require a production script. Add script text and resubmit the treatment.', {
-      status: 400, code: 'VALIDATION_ERROR',
-    });
-  }
-  const scenes = [...treatment.scenes].sort((a, b) => a.order - b.order);
-  const ids = new Set();
-  const orders = new Set();
-  let elapsed = 0;
-  const shots = scenes.map((scene) => {
-    validateVideoShot(project, scene, elapsed === 0);
-    if (ids.has(scene.sceneId) || orders.has(scene.order)) {
-      throw new ServerError('Video scenes must have unique sceneId and order values', { status: 400, code: 'VALIDATION_ERROR' });
-    }
-    ids.add(scene.sceneId);
-    orders.add(scene.order);
-    const startSeconds = elapsed;
-    elapsed = Math.round((elapsed + scene.durationSeconds) * 1000000) / 1000000;
-    return {
-      shotId: `shot-${scene.sceneId}`,
-      sceneId: scene.sceneId,
-      startSeconds,
-      endSeconds: elapsed,
-      durationSeconds: scene.durationSeconds,
-    };
-  });
-  if (Math.abs(elapsed - project.targetDurationSeconds) > 0.000001) {
-    throw new ServerError(`Video shots total ${elapsed}s; they must total the exact target of ${project.targetDurationSeconds}s`, { status: 400, code: 'VALIDATION_ERROR' });
-  }
-  const references = (project.videoPlanningContext?.references || project.videoDraft?.sources || []).map((source) => ({
-    ...source, referenceId: `${source.kind}:${source.id}`,
-    ...(sourceRevisions ? { sourceRevision: sourceRevisions[`${source.kind}:${source.id}`] ?? source.sourceRevision ?? null } : {}),
-  }));
-  if (new Set(references.map((ref) => ref.referenceId)).size !== references.length) {
-    throw new ServerError('Video source references must have unique kind and id values', { status: 400, code: 'VALIDATION_ERROR' });
-  }
-  return {
-    scriptId: project.treatment?.artifact?.scriptId || `script-${project.id}`,
-    revision: (project.treatment?.artifact?.revision || 0) + 1,
-    targetDurationSeconds: project.targetDurationSeconds,
-    aspectRatio: project.aspectRatio,
-    ...(project.videoPlanningContext ? { sourceContextRevision: project.videoPlanningContext.revision } : {}),
-    // Keep the selected IDs/revisions, never copy or mutate creative-suite records.
-    references,
-    shots,
-    stale: false,
-  };
-}
 
 function priorVideoTreatments(project) {
   if (!project.treatment?.artifact) return [];
